pyproject.toml
[tool.poetry]
name = "trainer"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.1.1"
google-api-python-client = "^1.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
